move performance calculation into usecases
